nmbd: Fixing CID 1509039 for time_t in DEBUG statement
Fix: typecast changing from (int)t to (intmax_t)t as intmax_t
can hold epoch seconds after 2038 year

+++ source3/libads/sitename_cache.c
@@ -63,8 +63,8 @@ bool sitename_store(const char *realm, const char *sitename)
 
 	expire = get_time_t_max(); /* Store indefinitely. */
 
-	DEBUG(10,("sitename_store: realm = [%s], sitename = [%s], expire = [%u]\n",
-		realm, sitename, (unsigned int)expire ));
+	DEBUG(10,("sitename_store: realm = [%s], sitename = [%s], expire = [%jd]\n",
+		realm, sitename, (intmax_t)expire ));
 
 	ret = gencache_set( key, sitename, expire );
 	TALLOC_FREE(key);
